Refactor BuyProduct and move Size/Qty Dropdowns to new file
