+++ src/components/ProductDetail/BuyProduct.jsx
@@ -2,48 +2,30 @@
 /* eslint-disable no-restricted-syntax */
 /* eslint-disable react/jsx-one-expression-per-line */
 /* eslint-disable react/forbid-prop-types */
-import React, { useState } from 'react';
+import React, {
+  useState,
+} from 'react';
 import PropTypes from 'prop-types';
 import {
-  Grid, Divider, Header, Breadcrumb, Dropdown, Icon, Button,
+  Grid, Divider, Header, Breadcrumb, Icon, Button,
 } from 'semantic-ui-react';
 import '../../App.css';
 import './BuyProduct.css';
 import PriceTag from './BuyProduct/PriceTag';
 import StyleThumbnails from './BuyProduct/StyleThumbnails';
+import SizeQtyDropDowns from './BuyProduct/SizeQtyDropdowns';
 import ShareButton from './BuyProduct/ShareButtons';
 
 function BuyProduct({
   essentials, currentStyle, styles, changeStyle,
 }) {
-  const qtyOptions = {};
-  const sizes = [];
-  const [currentQty, setCurrentQty] = useState([]);
+  const [isTrue, setIsTrue] = useState(true);
 
   if (currentStyle.length > 0) {
     const sections = [
       { key: 'Style', content: 'Style', link: false },
       { key: 'Stylename', content: currentStyle[0].name, link: false },
     ];
-
-    const isTrue = currentQty.length === 0;
-
-    for (const sku in currentStyle[0].skus) {
-      if (sku === 'null') {
-        break;
-      } else {
-        const item = currentStyle[0].skus[sku];
-        if (item.quantity > 0) {
-          const sizeOption = {
-            key: item.size,
-            text: item.size,
-            value: item.size,
-          };
-          sizes.push(sizeOption);
-          qtyOptions[item.size] = item.quantity;
-        }
-      }
-    }
 
     return (
       <div>
@@ -58,35 +40,7 @@ function BuyProduct({
         </Grid.Row>
         <Divider hidden />
         <Grid.Row>
-          <Dropdown
-            header="Select a size"
-            selection
-            onChange={(e) => {
-              const quantities = [];
-              let n = 1;
-              while (n <= qtyOptions[e.target.textContent]) {
-                if (n > 15) {
-                  break;
-                }
-                quantities.push({
-                  key: n,
-                  text: n,
-                  value: n,
-                });
-                n += 1;
-              }
-              setCurrentQty(quantities);
-            }}
-            options={sizes}
-            placeholder={sizes.length === 0 ? 'OUT OF STOCK' : 'Select Size'}
-          />
-          <Dropdown
-            placeholder={currentQty.length === 0 ? '-' : '1'}
-            compact
-            defaultValue="1"
-            selection
-            options={currentQty}
-          />
+          <SizeQtyDropDowns currentStyle={currentStyle} setIsTrue={setIsTrue} />
         </Grid.Row>
         <Divider hidden />
         <Grid.Row>
